refactor: extract render styles
